+++ crates/sui-cluster-test/src/faucet.rs
@@ -114,7 +114,7 @@ impl FaucetClient for LocalFaucetClient {
     async fn request_sui_coins(&self, request_address: SuiAddress) -> FaucetResponse {
         let receipt = self
             .simple_faucet
-            .send(Uuid::new_v4(), request_address, &[200000000; 5])
+            .send(Uuid::new_v4(), request_address, &[200_000_000_000; 5])
             .await
             .unwrap_or_else(|err| panic!("Failed to get gas tokens with error: {}", err));
